# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * MMO Shop API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  sourceId?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  isActive: boolean;
  inventoryCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface ProductInput {
  name: string;
  description?: string;
  price: number;
  category?: string;
  sourceId?: string;
  imageUrl?: string;
  isActive?: boolean;
}

export interface ProductUpdate {
  name?: string;
  description?: string;
  price?: number;
  category?: string;
  imageUrl?: string;
  isActive?: boolean;
}

export type InventoryItemStatus = typeof InventoryItemStatus[keyof typeof InventoryItemStatus];


export const InventoryItemStatus = {
  available: 'available',
  sold: 'sold',
  reserved: 'reserved',
} as const;

export interface InventoryItem {
  id: number;
  productId: number;
  content: string;
  status: InventoryItemStatus;
  createdAt: string;
  /** @nullable */
  soldAt?: string | null;
}

export interface InventoryInput {
  /** List of account strings / keys to add */
  items: string[];
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  pending_payment: 'pending_payment',
  paid: 'paid',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface Order {
  id: number;
  telegramUserId: number;
  /** @nullable */
  telegramUsername?: string | null;
  productId?: number;
  /** @nullable */
  productName?: string | null;
  quantity: number;
  totalPrice: number;
  status: OrderStatus;
  /** @nullable */
  paymentMethod?: string | null;
  /** @nullable */
  deliveredItems?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type OrderUpdateStatus = typeof OrderUpdateStatus[keyof typeof OrderUpdateStatus];


export const OrderUpdateStatus = {
  pending_payment: 'pending_payment',
  paid: 'paid',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface OrderUpdate {
  status?: OrderUpdateStatus;
  paymentMethod?: string;
}

export interface Stats {
  totalRevenue: number;
  pendingOrders: number;
  totalProducts: number;
  totalInventory: number;
  totalUsers: number;
  todayOrders: number;
}

export interface SyncResult {
  synced: number;
  message: string;
}

export interface BotUser {
  id: number;
  telegramId: number;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  lastName?: string | null;
  createdAt: string;
}

export interface RevenuePoint {
  date: string;
  revenue: number;
  orders: number;
}

export type AddInventoryItems201 = {
  added: number;
};

export type ListOrdersParams = {
status?: string;
limit?: number;
};